Add get_research_authors util function
