+++ network-api/networkapi/wagtailpages/pagemodels/profiles.py
@@ -8,9 +8,6 @@ from wagtail_localize.fields import SynchronizedField, TranslatableField
 
 
 class ProfileQuerySet(models.QuerySet):
-    def filter_research_authors(self):
-        return self.filter(authored_research__isnull=False).distinct()
-
     def filter_blog_authors(self):
         return self.filter(blogauthors__isnull=False).distinct()
